+++ RELEASE-NOTES.md
@@ -4,6 +4,10 @@ Release Notes
 0.3.0 (in progress)
 -------------------
 - using copy of data internally when creating streams to avoid ConcurrentModificationException if underlying set is changed during stream handling
+- added support for Java 8 Time API. Beneath java.util.date you may also use java.time.LocalDateTime in JSON entities
+- refactored package structure, com.github.christiangroth.jsonstore.JsonStore moved to com.github.christiangroth.jsonstore.store subpackage
+- added com.github.christiangroth.jsonstore.store.JsonSingletonStore to provide storage for single object only
+- fixed NullPointerException during initialization of transient JsonStores
 
 0.2.0
 -----